datapath/service: Introduce local redirect service
An entry of the type "local redirect" will be created when a
user installs a local redirect policy (LRP) to redirect traffic
from a frontend (L3 + L4 or K8s service cluster IP + port) tuple
to a backend within a node. The service thus will only have
node-local backends. The lifecycle of a local redirect service
will be tied to its associated LRP.

When an LRP enables traffic redirection for a K8s service, it's service
type will be changed to local redirect type, and the backends list will
be filtered to only include node-local backends.

Testing:
cilium service list
ID   Frontend               Service Type    Backend
12   169.254.169.254:8080   LocalRedirect   1 => 10.16.69.61:80

sudo cilium bpf lb list
SERVICE ADDRESS        BACKEND ADDRESS

169.254.169.254:8080   0.0.0.0:0 (12) [LocalRedirect]
                       10.16.69.61:80 (12)

kubectl exec nginx-client -- curl -I http://169.254.169.254:8080/index.html
HTTP/1.1 200 OK

cilium -D monitor --related-to 142
Listening for events on 6 CPUs with 64x4096 of shared memory
Press Ctrl-C to quit
level=info msg="Initializing dissection cache..." subsys=monitor
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp SYN
-> endpoint 142 flow 0x682a7114 identity 10591->7086 state reply ifindex lxce3c498b8259e orig-ip 10.16.69.61: 169.254.169.254:8080 -> 10.16.224.211:52718 tcp SYN, ACK
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp ACK
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp ACK
-> endpoint 142 flow 0x682a7114 identity 10591->7086 state reply ifindex lxce3c498b8259e orig-ip 10.16.69.61: 169.254.169.254:8080 -> 10.16.224.211:52718 tcp ACK
-> endpoint 142 flow 0x682a7114 identity 10591->7086 state reply ifindex lxce3c498b8259e orig-ip 10.16.69.61: 169.254.169.254:8080 -> 10.16.224.211:52718 tcp ACK
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp ACK
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp ACK, FIN
-> endpoint 142 flow 0x682a7114 identity 10591->7086 state reply ifindex lxce3c498b8259e orig-ip 10.16.69.61: 169.254.169.254:8080 -> 10.16.224.211:52718 tcp ACK, FIN
<- endpoint 142 flow 0x480114f0 identity 7086->0 state new ifindex 0 orig-ip 0.0.0.0: 10.16.224.211:52718 -> 169.254.169.254:8080 tcp ACK

+++ pkg/maps/lbmap/lbmap.go
@@ -470,7 +470,7 @@ func updateMasterService(fe ServiceKey, nbackends int, revNATID int, svcType loa
 		IsRoutable:       !fe.IsSurrogate(),
 		CheckSourceRange: checkSourceRange,
 	})
-	zeroValue.SetFlags(flag.UInt8())
+	zeroValue.SetFlags(flag.UInt16())
 	if sessionAffinity {
 		zeroValue.SetSessionAffinityTimeoutSec(sessionAffinityTimeoutSec)
 	}
